+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleDataprocService.scala
@@ -19,7 +19,7 @@ import org.broadinstitute.dsde.workbench.model.google.{GcsBucketName, GoogleProj
 import org.typelevel.log4cats.StructuredLogger
 
 import java.util.concurrent.ScheduledThreadPoolExecutor
-import scala.collection.JavaConverters._
+import scala.jdk.CollectionConverters._
 
 /**
  * Algebra for Google Dataproc access