+++ project/Settings.scala
@@ -6,7 +6,6 @@ import sbt.Keys.{scalacOptions, _}
 import sbt._
 import scoverage.ScoverageKeys.coverageExcludedPackages
 
-//noinspection TypeAnnotation
 object Settings {
 
   val artifactory = "https://broadinstitute.jfrog.io/broadinstitute/"
@@ -27,50 +26,6 @@ object Settings {
   )
 
   lazy val commonCompilerSettings = scalacOptions ++= (CrossVersion.partialVersion(scalaVersion.value) match {
-    case Some((2, 12)) =>
-      Seq(
-        "-deprecation", // Emit warning and location for usages of deprecated APIs.
-        "-encoding",
-        "utf-8", // Specify character encoding used by source files.
-        "-explaintypes", // Explain type errors in more detail.
-        "-feature", // Emit warning and location for usages of features that should be imported explicitly.
-        "-language:existentials", // Existential types (besides wildcard types) can be written and inferred
-        "-language:higherKinds", // Allow higher-kinded types
-        "-language:implicitConversions", // Allow definition of implicit functions called views
-        "-unchecked", // Enable additional warnings where generated code depends on assumptions.
-//        "-Xcheckinit", // Wrap field accessors to throw an exception on uninitialized access.
-//        "-Xfatal-warnings", // Fail the compilation if there are any warnings.
-        "-Xfuture", // Turn on future language features.
-        "-Xlint:adapted-args", // Warn if an argument list is modified to match the receiver.
-        "-Xlint:by-name-right-associative", // By-name parameter of right associative operator.
-        "-Xlint:constant", // Evaluation of a constant arithmetic expression results in an error.
-        "-Xlint:delayedinit-select", // Selecting member of DelayedInit.
-        "-Xlint:doc-detached", // A Scaladoc comment appears to be detached from its element.
-        "-Xlint:inaccessible", // Warn about inaccessible types in method signatures.
-        "-Xlint:infer-any", // Warn when a type argument is inferred to be `Any`.
-        "-Xlint:missing-interpolator", // A string literal appears to be missing an interpolator id.
-        "-Xlint:nullary-override", // Warn when non-nullary `def f()' overrides nullary `def f'.
-        "-Xlint:nullary-unit", // Warn when nullary methods return Unit.
-        "-Xlint:option-implicit", // Option.apply used implicit view.
-        "-Xlint:package-object-classes", // Class or object defined in package object.
-        "-Xlint:poly-implicit-overload", // Parameterized overloaded implicit methods are not visible as view bounds.
-        "-Xlint:private-shadow", // A private field (or class parameter) shadows a superclass field.
-        "-Xlint:stars-align", // Pattern sequence wildcard must align with sequence component.
-        "-Xlint:type-parameter-shadow", // A local type parameter shadows a type already in scope.
-        "-Xlint:unsound-match", // Pattern match may not be typesafe.
-        "-Yno-adapted-args", // Do not adapt an argument list (either by inserting () or creating a tuple) to match the receiver.
-        "-Ypartial-unification", // Enable partial unification in type constructor inference
-        "-Ywarn-dead-code", // Warn when dead code is identified.
-        "-Ywarn-extra-implicit", // Warn when more than one implicit parameter section is defined.
-        "-Ywarn-inaccessible", // Warn about inaccessible types in method signatures.
-        "-Ywarn-infer-any", // Warn when a type argument is inferred to be `Any`.
-        "-Ywarn-nullary-override", // Warn when non-nullary `def f()' overrides nullary `def f'.
-        "-Ywarn-nullary-unit", // Warn when nullary methods return Unit.
-        "-Ywarn-unused:implicits", // Warn if an implicit parameter is unused.
-        "-Ywarn-unused:imports", // Warn if an import selector is not referenced.
-        "-Ywarn-unused:patvars", // Warn if a variable bound in a pattern is unused.
-        "-language:postfixOps"
-      )
     case Some((2, 13)) =>
       Seq(
 //      "-deprecation", // Emit warning and location for usages of deprecated APIs. TODO: enable this when we migrate off of 2.12
@@ -112,8 +67,8 @@ object Settings {
   })
 
   val scala213 = "2.13.8"
-  val cross212and213 = Seq(
-    crossScalaVersions := List("2.12.15", scala213)
+  val cross212and3 = Seq(
+    crossScalaVersions := List(scala213, "3.1.2")
   )
 
   // common settings for all sbt subprojects
@@ -124,38 +79,38 @@ object Settings {
     commonCompilerSettings
   )
 
-  val utilSettings = cross212and213 ++ commonSettings ++ List(
+  val utilSettings = commonSettings ++ List(
     name := "workbench-util",
     libraryDependencies ++= utilDependencies,
     version := createVersion("0.6")
   ) ++ publishSettings
 
-  val util2Settings = cross212and213 ++ commonSettings ++ List(
+  val util2Settings = commonSettings ++ List(
     name := "workbench-util2",
     libraryDependencies ++= util2Dependencies,
     version := createVersion("0.2")
   ) ++ publishSettings
 
-  val modelSettings = cross212and213 ++ commonSettings ++ List(
+  val modelSettings = commonSettings ++ List(
     name := "workbench-model",
     libraryDependencies ++= modelDependencies,
     version := createVersion("0.15")
   ) ++ publishSettings
 
-  val metricsSettings = cross212and213 ++ commonSettings ++ List(
+  val metricsSettings = commonSettings ++ List(
     name := "workbench-metrics",
     libraryDependencies ++= metricsDependencies,
     version := createVersion("0.5")
   ) ++ publishSettings
 
-  val googleSettings = cross212and213 ++ commonSettings ++ List(
+  val googleSettings = commonSettings ++ List(
     name := "workbench-google",
     libraryDependencies ++= googleDependencies,
     version := createVersion("0.21"),
     coverageExcludedPackages := ".*HttpGoogle.*DAO.*"
   ) ++ publishSettings
 
-  val google2Settings = cross212and213 ++ commonSettings ++ List(
+  val google2Settings = commonSettings ++ List(
     name := "workbench-google2",
     libraryDependencies ++= google2Dependencies,
     version := createVersion("0.24")
@@ -183,6 +138,12 @@ object Settings {
     name := "workbench-notifications",
     libraryDependencies ++= notificationsDependencies,
     version := createVersion("0.3")
+  ) ++ publishSettings
+
+  val oauth2Settings = commonSettings ++ List(
+    name := "workbench-oauth2",
+    libraryDependencies ++= oauth2Depdendencies,
+    version := createVersion("0.1")
   ) ++ publishSettings
 
   val rootSettings = commonSettings ++ noPublishSettings ++ noTestSettings