+++ google2/CHANGELOG.md
@@ -7,6 +7,7 @@ This file documents changes to the `workbench-google2` library, including notes 
 Changed:
 - `GoogleComputeService[F]`'s APIs are updated for `google-cloud-compute` upgrade
 - Remove `ComputePollOperation`
+- Drop 2.12 support
 
 Dependency Upgrades:
 | Dependency   |      Old Version      |  New Version |
@@ -17,12 +18,14 @@ Dependency Upgrades:
 | com.google.guava |   31.0.1-jre   |   31.1-jre |
 | google-cloud-nio |   0.123.25   |   0.123.27 |
 | google-cloud-container |   2.3.0   |   2.3.6 |
+| google-cloud-firestore | 2.6.2 | 3.0.21 |
+| google-cloud-storage-transfer | 0.2.2 | 1.0.4 |
 | client-java |   14.0.1   |   15.0.0 |
 | google-cloud-resourcemanager |   1.2.0   |  1.2.7 |
 | simpleclient_httpserver |   0.12.0   | 0.15.0 |
 | log4cats-slf4j |  2.1.1   |  2.2.0 |
 
-SBT Dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.24-311c84b"`
+SBT Dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.24-447afa5"`
 
 ## 0.23
 Added: